+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 - Adjusted serialization format to be more compact
 - Re-select task after successfully editing tags
 - Use Unicode aware lowercase in task summary search
+- Fixed reverse task search not progressing under certain conditions
 - Bumped minimum supported Rust version to `1.52`
 - Bumped `dirs` dependency to `4.0`